requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
cachetools==5.3.2 